# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API example laravel 8 application",
        "version": "1.0"
    },
    "paths": {
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get list of categories or one category by ID of category",
                "description": "Get category or category by ID",
                "operationId": "f41c4b9deeaf40fb72f3521d6b2275bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "integer",
                            "nullable": true
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get category or category"
                    }
                }
            },
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "Update category by ID of category",
                "description": "Update category by ID of category",
                "operationId": "ccde27a4109998e78956003f0a234670",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "response message - true / false",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User is unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "categories"
                ],
                "summary": "Delete category by ID of category",
                "description": "Delete category by ID",
                "operationId": "844745393ab4e68658c5523e664e3005",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete category and get response message - true / false",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User is unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/categories/": {
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Create new category",
                "description": "Create new category",
                "operationId": "455a9916673c41bec23ce7447473a664",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "response message - true / false",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User is unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get list of products or one product by ID of product",
                "description": "Get products or product by ID",
                "operationId": "c0edca2ead60165659257ddf151d24f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "integer",
                            "nullable": true
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get products or product"
                    }
                }
            },
            "put": {
                "tags": [
                    "products"
                ],
                "summary": "Update product by ID of product",
                "description": "Update product by ID of product",
                "operationId": "3f51692dbf9ef513f6b245036979d9eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "response message - true / false",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User is unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Delete product by ID of product",
                "description": "Delete product by ID",
                "operationId": "cf9a9d6febcfb1dc188b07eafcc48c7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete product and get response message - true / false",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User is unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/": {
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Create new product",
                "description": "Create new product",
                "operationId": "10a5bcb429b376de5e57c761459e4102",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "response message - true / false",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User is unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CategoryRequest": {
                "properties": {
                    "name": {
                        "description": "category name",
                        "format": "string",
                        "default": "lorem ipsum"
                    },
                    "description": {
                        "description": "category description",
                        "format": "string",
                        "default": "description of lorem ipsum"
                    }
                },
                "type": "object"
            },
            "CategoryUpdateRequest": {
                "properties": {
                    "name": {
                        "description": "category name",
                        "format": "string",
                        "default": "lorem ipsum"
                    },
                    "description": {
                        "description": "category description",
                        "format": "string",
                        "default": "description of lorem ipsum"
                    },
                    "active": {
                        "description": "is categoy active",
                        "type": "boolean",
                        "default": true
                    },
                    "deleted": {
                        "description": "is categoy deleted",
                        "type": "boolean",
                        "default": false
                    }
                },
                "type": "object"
            },
            "ProductRequest": {
                "properties": {
                    "name": {
                        "description": "product name",
                        "format": "string",
                        "default": "lorem ipsum dolor"
                    },
                    "description": {
                        "description": "product description",
                        "format": "string",
                        "default": "description of lorem ipsum dolor"
                    },
                    "category_id": {
                        "description": "category of product",
                        "format": "integer",
                        "default": "1"
                    },
                    "file": {
                        "description": "main photo of product",
                        "type": "file"
                    }
                },
                "type": "object"
            },
            "ProductUpdateRequest": {
                "properties": {
                    "name": {
                        "description": "product name",
                        "format": "string",
                        "default": "lorem ipsum dolor"
                    },
                    "description": {
                        "description": "product description",
                        "format": "string",
                        "default": "description of lorem ipsum dolor"
                    },
                    "category_id": {
                        "description": "category of product",
                        "format": "integer",
                        "default": "1"
                    },
                    "file": {
                        "description": "main photo of product",
                        "type": "file"
                    },
                    "active": {
                        "description": "is product active",
                        "type": "boolean",
                        "default": true
                    },
                    "deleted": {
                        "description": "is product deleted",
                        "type": "boolean",
                        "default": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "barerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}